+++ packages/db/src/index.ts
@@ -12,6 +12,7 @@ export {
   fakeGranuleRecordFactory,
   fakePdrRecordFactory,
   fakeProviderRecordFactory,
+  fakeReconciliationReportRecordFactory,
   fakeRuleRecordFactory,
   generateLocalTestDb,
 } from './test-utils';
@@ -31,7 +32,6 @@ export {
 export {
   BaseRecord,
 } from './types/base';
-
 export {
   PostgresAsyncOperation,
   PostgresAsyncOperationRecord,
@@ -45,56 +45,50 @@ export {
   PostgresExecutionRecord,
 } from './types/execution';
 export {
-  PostgresProvider,
-  PostgresProviderRecord,
-} from './types/provider';
-export {
-  PostgresRule,
-  PostgresRuleRecord,
-} from './types/rule';
+  PostgresFile,
+  PostgresFileRecord,
+} from './types/file';
 export {
   PostgresGranule,
   PostgresGranuleRecord,
 } from './types/granule';
 export {
+  PostgresGranuleExecution,
+} from './types/granule-execution';
+export {
   PostgresPdr,
   PostgresPdrRecord,
 } from './types/pdr';
 export {
-  PostgresFile,
-  PostgresFileRecord,
-} from './types/file';
+  PostgresProvider,
+  PostgresProviderRecord,
+} from './types/provider';
 export {
-  PostgresGranuleExecution,
-} from './types/granule-execution';
+  PostgresReconciliationReport,
+  PostgresReconciliationReportRecord,
+} from './types/reconciliation_report';
+export {
+  PostgresRule,
+  PostgresRuleRecord,
+} from './types/rule';
+
 export {
   translateApiAsyncOperationToPostgresAsyncOperation,
   translatePostgresAsyncOperationToApiAsyncOperation,
 } from './translate/async_operations';
 export {
-  translateApiFiletoPostgresFile,
-  translatePostgresFileToApiFile,
-} from './translate/file';
-
-export {
   translateApiCollectionToPostgresCollection,
   translatePostgresCollectionToApiCollection,
 } from './translate/collections';
-
-export {
-  translateApiProviderToPostgresProvider,
-  translatePostgresProviderToApiProvider,
-} from './translate/providers';
-export {
-  translatePostgresRuleToApiRule,
-  translateApiRuleToPostgresRule,
-  translateApiRuleToPostgresRuleRaw,
-} from './translate/rules';
 export {
   translateApiExecutionToPostgresExecution,
   translateApiExecutionToPostgresExecutionWithoutNilsRemoved,
   translatePostgresExecutionToApiExecution,
 } from './translate/executions';
+export {
+  translateApiFiletoPostgresFile,
+  translatePostgresFileToApiFile,
+} from './translate/file';
 export {
   translateApiGranuleToPostgresGranule,
   translateApiGranuleToPostgresGranuleWithoutNilsRemoved,
@@ -105,6 +99,18 @@ export {
   translateApiPdrToPostgresPdr,
   translatePostgresPdrToApiPdr,
 } from './translate/pdr';
+export {
+  translateApiProviderToPostgresProvider,
+  translatePostgresProviderToApiProvider,
+} from './translate/providers';
+export {
+  translatePostgresReconciliationReportToApiReconciliationReport,
+} from './translate/reconciliation_reports';
+export {
+  translatePostgresRuleToApiRule,
+  translateApiRuleToPostgresRule,
+  translateApiRuleToPostgresRuleRaw,
+} from './translate/rules';
 
 export {
   getCollectionsByGranuleIds,
@@ -140,8 +146,11 @@ export {
   QuerySearchClient,
 } from './lib/QuerySearchClient';
 export {
-  BaseSearch,
-} from './search/BaseSearch';
+  AsyncOperationSearch,
+} from './search/AsyncOperationSearch';
+export {
+  CollectionSearch,
+} from './search/CollectionSearch';
 export {
   ExecutionSearch,
 } from './search/ExecutionSearch';
@@ -149,14 +158,17 @@ export {
   GranuleSearch,
 } from './search/GranuleSearch';
 export {
+  ProviderSearch,
+} from './search/ProviderSearch';
+export {
+  RuleSearch,
+} from './search/RuleSearch';
+export {
   StatsSearch,
 } from './search/StatsSearch';
-export {
-  CollectionSearch,
-} from './search/CollectionSearch';
 
-export { AsyncOperationPgModel } from './models/async_operation';
 export { BasePgModel } from './models/base';
+export { AsyncOperationPgModel } from './models/async_operation';
 export { CollectionPgModel } from './models/collection';
 export { ExecutionPgModel } from './models/execution';
 export { FilePgModel } from './models/file';
@@ -164,4 +176,5 @@ export { GranulePgModel } from './models/granule';
 export { GranulesExecutionsPgModel } from './models/granules-executions';
 export { PdrPgModel } from './models/pdr';
 export { ProviderPgModel } from './models/provider';
+export { ReconciliationReportPgModel } from './models/reconciliation_report';
 export { RulePgModel } from './models/rule';